Add a new type `RawType`
This is a generic type that allows the user to specific
an injectable or instance that is registered in the container to be
pass in. Not every type has to be wrapped in `Service<T>` and instances
of these types must be injectable into a function or closure pass to
`inject_and_call`. This type makes it easy...